[SPRINT-1][BP-27] Plant_Identification_Features Create HealthAssessment functionality with its respective needed classes. Separated DrPlantaController.java into other controllers Created PaginationUtils.java for pagination centralization
