+++ src/main/java/com/blooming/api/response/dto/HealthAssessmentDTO.java
@@ -1,0 +1,20 @@
+package com.blooming.api.response.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.List;
+
+@Setter
+@Getter
+public class HealthAssessmentDTO {
+
+    private List<DiseaseSuggestionDTO> diseaseSuggestions;
+
+    @Override
+    public String toString() {
+        return "HealthAssessmentDTO{" +
+                "diseaseSuggestions=" + diseaseSuggestions +
+                '}';
+    }
+}
