+++ src/main/java/com/blooming/api/service/plantAI/IPlantAIService.java
@@ -1,6 +1,7 @@
 package com.blooming.api.service.plantAI;
 
 import com.blooming.api.entity.PlantIdentified;
+import com.blooming.api.response.dto.HealthAssessmentDTO;
 import com.blooming.api.response.dto.WateringDayDTO;
 import com.blooming.api.response.dto.PlantSuggestionDTO;
 
@@ -9,6 +10,8 @@ import java.util.List;
 public interface IPlantAIService {
 
     List<PlantSuggestionDTO> identifyImage(byte[] bytes);
+
+    List<HealthAssessmentDTO> generateHealthAssessment(byte[] bytesImg);
 
     PlantIdentified getPlantInformationByName(String plantName, String tokenPlant);
