+++ src/main/resources/application.properties
@@ -15,6 +15,8 @@ security.jwt.expiration-time=3600000
 # Plant ID
 plant.id.api.identify.url=https://plant.id/api/v3/identification
 plant.id.api.plant.search.url=https://plant.id/api/v3/kb/plants/name_search?q=
+plant.id.api.plant.health.url=https://plant.id/api/v3/health_assessment?details=local_name,description,url,treatment
 plant.id.api.base.url=https://plant.id/api/v3/kb/plants/
 plant.id.api.key=
+
